Add file input retriever class for reading from file source (#86)
* add file input retriever

* move input source check to inputs.py

* create unittests for file input retriever
