+++ genai-perf/genai_perf/inputs/inputs.py
@@ -61,16 +61,11 @@ class Inputs:
         return json_in_pa_format
 
     def _check_for_valid_args(self) -> None:
+        self._check_for_supported_input_type()
         self._check_for_dataset_name_if_input_type_is_url()
         self._check_for_tokenzier_if_input_type_is_synthetic()
         self._check_for_valid_starting_index()
         self._check_for_valid_length()
-
-    def _verify_file(self) -> None:
-        if not self.config.input_filename.exists():
-            raise FileNotFoundError(
-                f"The file '{self.config.input_filename}' does not exist."
-            )
 
     def _convert_generic_json_to_output_format(self, generic_dataset) -> Dict:
         converter = OutputFormatConverterFactory.create(self.config.output_format)
@@ -80,6 +75,18 @@ class Inputs:
         filename = self.config.output_dir / DEFAULT_INPUT_DATA_JSON
         with open(str(filename), "w") as f:
             f.write(json.dumps(json_in_pa_format, indent=2))
+
+    def _check_for_supported_input_type(self) -> None:
+        if self.config.output_format in [
+            OutputFormat.OPENAI_EMBEDDINGS,
+            OutputFormat.RANKINGS,
+            OutputFormat.IMAGE_RETRIEVAL,
+        ]:
+            if self.config.input_type != PromptSource.FILE:
+                raise GenAIPerfException(
+                    f"{self.config.output_format.to_lowercase()} only supports "
+                    "a file as input source."
+                )
 
     def _check_for_dataset_name_if_input_type_is_url(self) -> None:
         if (
