+++ genai-perf/tests/test_input_image_retrieval.py
@@ -27,7 +27,7 @@ class TestInputsImageRetrieval:
         return_value="data:image/jpeg;base64,/9j/4AAQSkZJRgABAQAAAQABAAD/2wBDAAgGBgcGBQgHBwcJCQgKDBQNDAsLDBkSEw8UHRofHh0aHBwgJC4nICIsIxwcKDcpLDAxNDQ0Hyc5PTgyPC4zNDL/",
     )
     @patch(
-        "genai_perf.inputs.input_retriever_factory.InputRetrieverFactory._get_input_dataset_from_file"
+        "genai_perf.inputs.input_retriever_factory.FileInputRetriever._get_input_dataset_from_file"
     )
     def test_image_retrieval(self, mock_get_input, mock_encode_image, mock_write_json):
         mock_get_input.return_value = {
@@ -80,7 +80,7 @@ class TestInputsImageRetrieval:
 
     @patch("genai_perf.inputs.inputs.Inputs._write_json_to_file")
     @patch(
-        "genai_perf.inputs.input_retriever_factory.InputRetrieverFactory._get_input_dataset_from_file"
+        "genai_perf.inputs.input_retriever_factory.FileInputRetriever._get_input_dataset_from_file"
     )
     @patch(
         "genai_perf.inputs.input_retriever_factory.InputRetrieverFactory._encode_image"
